refactor(tests): migrate mobrender module #507
