+++ client/mobrender/components/block-change-background.spec.js
