+++ client/mobrender/components/block-config-menu.spec.js
