+++ client/mobrender/redux/action-creators/async-fetch-blocks.spec.js
@@ -3,33 +3,36 @@ import axios from 'axios'
 import MockAdapter from 'axios-mock-adapter'
 import configureStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
-import { asyncFetchWidgets } from '~client/mobrender/redux/action-creators'
+import { asyncFetchBlocks } from '~client/mobrender/redux/action-creators'
 import { createAction } from '~client/mobrender/redux/action-creators/create-action'
 import * as t from '~client/mobrender/redux/action-types'
 
-import rootReducer from '../mock-reducers/root-reducer'
+import rootReducer from './mock-reducers/root-reducer'
 
 // Mock axios
 const mockAxios = new MockAdapter(axios)
 const mobilizationId = 1
 const data = [
-  { id: 1, kind: 'draft' },
-  { id: 2, kind: 'draft' }
+  { id: 1, name: 'Lorem', position: 1 },
+  { id: 2, name: 'Ipsum', position: 2 },
+  { id: 3, name: 'Dolor', position: 3 }
 ]
-mockAxios.onGet(`/mobilizations/${mobilizationId}/widgets`).reply(200, data)
+mockAxios.onGet(
+  `/mobilizations/${mobilizationId}/blocks`
+).reply(200, data)
 
 // Mock store
 const store = configureStore(
   [thunk.withExtraArgument({ api: axios })]
 )(rootReducer)
 
-describe('client/morender/async-creators/async-fetch-widget', () => {
-  it('should dispatch actions to fetching widgets', () => {
+describe('~client/mobrender/redux/action-creators/async-fetch-blocks', () => {
+  it('should dispatch actions to move up block', () => {
     const expectedActions = [
-      createAction(t.FETCH_WIDGETS_REQUEST),
-      createAction(t.FETCH_WIDGETS_SUCCESS, data)
+      createAction(t.FETCH_BLOCKS_REQUEST),
+      createAction(t.FETCH_BLOCKS_SUCCESS, data)
     ]
-    return store.dispatch(asyncFetchWidgets(mobilizationId))
+    return store.dispatch(asyncFetchBlocks(mobilizationId))
       .then(() => {
         expect(store.getActions().length).to.equal(2)
         expect(store.getActions()[0]).to.deep.equal(expectedActions[0])
