+++ client/mobrender/redux/action-creators/async-move-down.spec.js
@@ -4,11 +4,11 @@ import axios from 'axios'
 import MockAdapter from 'axios-mock-adapter'
 import configureStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
-import { asyncDestroyBlock } from '~client/mobrender/redux/action-creators'
+import { asyncMoveDown } from '~client/mobrender/redux/action-creators'
 import { createAction } from '~client/mobrender/redux/action-creators/create-action'
 import * as t from '~client/mobrender/redux/action-types'
 
-import rootReducer from '../mock-reducers/root-reducer'
+import rootReducer from './mock-reducers/root-reducer'
 
 // Mock axios
 const mockAxios = new MockAdapter(axios)
@@ -17,11 +17,9 @@ const data = [
   { id: 2, name: 'Ipsum', position: 2 },
   { id: 3, name: 'Dolor', position: 3 }
 ]
-const mobilization = { id: 1 }
-const block = data[0]
-
-mockAxios.onDelete(
-  `/mobilizations/${mobilization.id}/blocks/${block.id}`,
+const block = {...data[1], position: 3}
+mockAxios.onPut(
+  `/mobilizations/1/blocks/2`, { block }
 ).reply(200, block)
 
 // Mock store
@@ -30,24 +28,26 @@ const store = configureStore(
 )(fromJS(rootReducer).mergeDeep({
   mobilizations: {
     list: {
-      data: [ mobilization ],
-      currentId: mobilization.id
+      data: [ { id: 1, name: 'Mob' } ],
+      currentId: 1
     },
     blocks: { data }
   }
 }).toJS())
 
-describe('~client/mobrender/redux/action-creators/async-destroy-block', () => {
-  it('should dispatch actions to destroy block', () => {
+describe('~client/mobrender/redux/action-creators/async-move-down', () => {
+  it('should dispatch actions to move up block', () => {
     const expectedActions = [
-      createAction(t.DESTROY_BLOCK_REQUEST),
-      createAction(t.DESTROY_BLOCK_SUCCESS, block)
+      createAction(t.UPDATE_BLOCK_REQUEST),
+      createAction(t.UPDATE_BLOCK_SUCCESS, block),
+      createAction(t.MOVE_BLOCK_DOWN, block)
     ]
-    return store.dispatch(asyncDestroyBlock(block))
+    return store.dispatch(asyncMoveDown(data[1]))
       .then(() => {
-        expect(store.getActions().length).to.equal(2)
+        expect(store.getActions().length).to.equal(3)
         expect(store.getActions()[0]).to.deep.equal(expectedActions[0])
         expect(store.getActions()[1]).to.deep.equal(expectedActions[1])
+        expect(store.getActions()[2]).to.deep.equal(expectedActions[2])
       })
   })
 })
