+++ client/mobrender/redux/action-creators/non-async.spec.js
