+++ client/mobrender/redux/reducers/edition.spec.js
