+++ client/mobrender/redux/reducers/mobilizations.spec.js
