+++ client/mobrender/redux/reducers/widgets.spec.js
