+++ client/mobrender/redux/selectors.spec.js
