+++ client/mobrender/widgets/draft/components/draft-button.spec.js
@@ -3,7 +3,7 @@ import { expect } from 'chai'
 import { mount } from 'enzyme'
 import { DraftButton } from '~client/mobrender/widgets/draft/components'
 
-describe('mobrender/widgets/draft/components/draft-button', () => {
+describe('client/mobrender/widgets/draft/components/draft-button', () => {
   const props = {
     icon: 'text',
     label: 'Texto',
